requirements.txt
argparse
pandas
scikit-learn
numpy
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.2.1+cu118
munch
tqdm
PyYAML
transformers